Scripts/instagram_bot/requirements.txt
instaloader>=4.13.1
anthropic>=0.27.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
APScheduler>=3.10.4
pytz>=2024.1
